requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
browser-cookie3==0.20.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
google-auth==2.40.1
google-genai==1.15.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
loguru==0.7.3
lz4==4.4.4
multidict==6.4.3
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
setuptools==80.4.0
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
